requirements.txt
streamlit==1.20.0
tensorflow==2.11.0
numpy==1.23.5
Pillow==9.4.0
streamlit-drawable-canvas==0.9.3
protobuf==3.19.6
